+++ src/main/java/net/berserker_rpg/item/weapons/WeaponsRegister.java
@@ -2,6 +2,7 @@ package net.berserker_rpg.item.weapons;
 
 import net.berserker_rpg.BerserkerClassMod;
 import net.berserker_rpg.item.BerserkerGroup;
+import net.berserker_rpg.spell.BerserkerSpells;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.item.Item;
 import net.minecraft.item.Items;
@@ -54,7 +55,10 @@ public class WeaponsRegister {
 
     //BERSERKER-AXE
     private static Weapon.Entry berserker_axes(String name, Weapon.CustomMaterial material, float damage) {
-        return entry(name, material, SpellSwordItem::new, new WeaponConfig(damage, berserker_axe_attackSpeed), Equipment.WeaponType.SWORD);
+        return entry(name, material, SpellSwordItem::new, new WeaponConfig(damage, berserker_axe_attackSpeed), Equipment.WeaponType.DOUBLE_AXE);
+    }
+    private static Weapon.Entry sword(String name, Weapon.CustomMaterial material, float damage) {
+        return entry(name, material, SpellSwordItem::new, new WeaponConfig(damage, -2.4f), Equipment.WeaponType.SWORD);
     }
 
     public static final Weapon.Entry flint_berserker_axe = berserker_axes("flint_berserker_axe",
@@ -84,6 +88,7 @@ public class WeaponsRegister {
     private static final String BETTER_END = "betterend";
     private static final String BETTER_NETHER = "betternether";
     private static final String AETHER = "aether";
+    private static final String ARSENAL = "arsenal";
     private static final String LNE = "loot_n_explore";
     private static final float lneWeaponSpellPower = 2.0F;
     private static final float lneAxeAttackDamage = 15.0F;
@@ -92,6 +97,7 @@ public class WeaponsRegister {
     public static Identifier avalanche = Identifier.of("loot_n_explore", "avalanche");
     public static Identifier waterbomb = Identifier.of("loot_n_explore", "waterbomb");
     public static Identifier wither_pulse = Identifier.of("loot_n_explore", "wither_pulse");
+    public static Identifier carve = Identifier.of("berserker_rpg", "carve");
     //Registration
     public static void register(Map<String, WeaponConfig> configs) {
         if(FabricLoader.getInstance().isModLoaded(BETTER_NETHER) || BerserkerClassMod.tweaksConfig.value.ignore_items_required_mods){
@@ -131,7 +137,7 @@ public class WeaponsRegister {
             berserker_axes( "wither_berserker_axe",
                     Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.BONE)),lneAxeAttackDamage)
                     .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
-                    .attribute(AttributeModifier.bonus(SpellSchools.ARCANE.id, lneWeaponSpellPower))
+                    .attribute(AttributeModifier.bonus(SpellSchools.SOUL.id, lneWeaponSpellPower))
                     .spell(wither_pulse)
                     .rarity = Rarity.RARE;
             berserker_axes( "glacial_berserker_axe",
@@ -140,9 +146,6 @@ public class WeaponsRegister {
                     .attribute(AttributeModifier.bonus(SpellSchools.FROST.id, lneWeaponSpellPower))
                     .spell(avalanche)
                     .rarity = Rarity.RARE;
-
-
-
 // ADD THIS IN LNE OR BERSERKER LNE MOD
             /*
             Identifier itemIdG = Identifier.of("loot_n_explore", "elder_guardian_axe");
@@ -159,6 +162,18 @@ public class WeaponsRegister {
                 content.addAfter(Registries.ITEM.get(itemIdG),Registries.ITEM.get(itemId3));
             });
             */
+        }
+        if (FabricLoader.getInstance().isModLoaded(ARSENAL) || BerserkerClassMod.tweaksConfig.value.ignore_items_required_mods) {
+            berserker_axes( "unique_berserker_axe_1",
+                    Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.IRON_BLOCK)),lneAxeAttackDamage)
+                    .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.10F))
+                    .spell(BerserkerSpells.carve_melee.id())
+                    .rarity = Rarity.RARE;
+            sword( "unique_sword_1",
+                    Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.IRON_BLOCK)),8.0F)
+                    .attribute(AttributeModifier.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:rage_modifier")),0.05F))
+                    .spell(BerserkerSpells.carve_melee.id())
+                    .rarity = Rarity.RARE;
         }
         Weapon.register(configs, entries, BerserkerGroup.BERSERKER_KEY);
     }